fix(ext-gateway): break server<->ext-proxy import cycle for canonical policy
The canonical gateway access policy and execution audit were owned by
server/ and re-exported into the @product/ext-proxy package, while server
already delegates to @product/ext-proxy/ee/handler. That closed a
server <-> @product/ext-proxy project cycle (failing the circular-deps
check and breaking nx's EE build DAG), and ee/server could not resolve
@product/ext-proxy/shared/* (failing typecheck) due to an erroneous ../
prefix in its path mapping.

- Move access.ts and executionAudit.ts into packages/product-ext-proxy/ee/gateway;
  server modules become thin re-export shims for existing importers.
- access.ts now uses @alga-psa/auth/rbac (behaviorally identical hasPermission)
  and a relative gateway-utils import, so the package no longer imports server.
- ext-proxy handler/auth consume the local copies; the ext-proxy -> server
  edge is removed and the cycle is gone.
- Fix ee/server tsconfig @product/ext-proxy/* mapping to be repo-root-relative
  like every sibling mapping.
- Update the two affected test mock paths accordingly.
