+++ ee/server/tsconfig.json
@@ -86,8 +86,8 @@
       "@product/chat/entry": ["packages/product-chat/oss/entry"],
       "@product/extensions/entry": ["packages/product-extensions/oss/entry"],
       "@product/extensions/*": ["packages/product-extensions/*"],
-      "@product/ext-proxy/handler": ["../packages/product-ext-proxy/ee/handler"],
-      "@product/ext-proxy/*": ["../packages/product-ext-proxy/*"],
+      "@product/ext-proxy/handler": ["packages/product-ext-proxy/ee/handler"],
+      "@product/ext-proxy/*": ["packages/product-ext-proxy/*"],
       "@product/settings-extensions/entry": ["packages/product-settings-extensions/oss/entry"],
       "@product/settings-extensions/*": ["packages/product-settings-extensions/*"],
       "@product/billing/entry": ["packages/product-billing/oss/entry"],
